Move dice package into dbnode/storage
